requirements.txt
openai-whisper
python-dotenv
openai
datasets